Implemented event rx/tx properly
Was able to finally get a good implementation to having the http web
server and led driver communicate back and forth with each other while
making the web server completely optional (i.e. a feature). The trick was
to provide the channels to the modules instead of having them construct
them. Also in the case of the led driver, wrap them in an Option to make
them optional. The implementation could still use a couple improvements
but current it is good enough to check in.