Using enum to static dispatch drawable
After a quick disucssion with the developer of embedded-layouts, he
confirmed that because Drawable has a generic in draw<D> there is no
current way in Rust to use dyn since building a vtable is impossible. To
get arround this we will use enum variants for static dispatch. I have
an example currently but the goal is to have a derive macro that builds
the impls.