Refactor code into person_tracker
We can now have multiple state providers for the person's state. This
means that multiple state providers will run in parallel and the most
useful will be rendered on the screen. This allows for easy integration
of new state providers and cleans up the code. Still have to write the
code the will glue this all together but the components are there.