Added derive(RenderFactory) enum
So assoicate types are not going to work in a dynamic way. There might
be some crazy way with macros, but I am not aware. Instead we will pass
a Reader into the Factory and that will read in the configuration file.
The downside of this is that we need to tell the render factory which
configuration format we are using, unlike the previous implementation
where the Factory did not care about the format.